fix: Use BigInt instead of int for the supply value as int is not big enough (should be uint64)

+++ packages/solana/lib/src/spl_token/spl_token.dart
@@ -32,7 +32,7 @@ class SplToken {
     final supplyValue = supplyResponse.value;
     return SplToken._(
       decimals: supplyValue.decimals,
-      supply: int.parse(supplyValue.amount),
+      supply: BigInt.parse(supplyValue.amount),
       rpcClient: rpcClient,
       mint: mint,
       owner: owner,
@@ -214,7 +214,7 @@ class SplToken {
   }
 
   final int decimals;
-  final int supply;
+  final BigInt supply;
   final String mint;
   final Ed25519HDKeyPair? owner;
   final RPCClient _rpcClient;
